+++ .github/workflows/checks.yml
@@ -43,11 +43,11 @@ jobs:
         run: cd snarkxt/ && cargo test
 
 
-      - name: Run format check for `blender-cli`
-        run: cd blender-cli/ && cargo fmt --all --check
+      - name: Run format check for `shielder-cli`
+        run: cd shielder-cli/ && cargo fmt --all --check
 
-      - name: Run clippy for `blender-cli`
-        run: cd blender-cli/ && cargo clippy --all-features -- --no-deps -D warnings
+      - name: Run clippy for `shielder-cli`
+        run: cd shielder-cli/ && cargo clippy --all-features -- --no-deps -D warnings
 
-      - name: Run tests  for `blender-cli`
-        run: cd blender-cli/ && cargo test
+      - name: Run tests  for `shielder-cli`
+        run: cd shielder-cli/ && cargo test
