+++ house-snark/src/lib.rs
@@ -4,11 +4,11 @@ pub mod serialization;
 
 pub use environment::{CircuitField, NonUniversalProvingSystem, RawKeys, SomeProvingSystem};
 #[cfg(feature = "deposit")]
-pub use relations::blender::DepositRelation;
+pub use relations::shielder::DepositRelation;
 #[cfg(feature = "withdraw")]
-pub use relations::blender::WithdrawRelation;
+pub use relations::shielder::WithdrawRelation;
 #[cfg(any(feature = "deposit", feature = "withdraw"))]
-pub use relations::blender::{
+pub use relations::shielder::{
     bytes_from_note, compute_note, note_from_bytes, MerklePath, MerkleRoot, Note, Nullifier,
     TokenAmount, TokenId, Trapdoor,
 };
