+++ house-snark/src/relations/shielder/note.rs
