+++ shielder-cli/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "blender-cli"
+name = "shielder-cli"
 version = "0.1.0"
 edition = "2021"
 authors = ["Cardinal Cryptography"]
 
 [dependencies]
-aleph_client = { git = "https://github.com/Cardinal-Cryptography/aleph-node/", branch = "blender-cli" }
+aleph_client = { git = "https://github.com/Cardinal-Cryptography/aleph-node/", branch = "shielder-cli" }
 anyhow = "1.0"
 chacha20poly1305 = { version = "0.10.1", features = ["stream"] }
 clap = { version = "4.0", features = ["derive"] }
