+++ shielder-cli/src/app_state.rs
