+++ shielder-cli/src/state_file.rs
